# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /ctp/jobs/escalate:
    post:
      operationId: escalateCtpConcerns
      summary: Dedicated scheduler credential only, never a staff or trusted-person session
      responses:
        "200": { description: Sample concern escalation sweep }
        "403": { description: Missing or invalid scheduler credential }
...
  /care-plans/generate:
    post:
      operationId: generateCarePlan
      tags: [care-plans]
      summary: Generate a structured draft care plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateCarePlanRequest"
      responses:
        "200":
          description: Generated draft
          content:
            application/json:
              schema:
                type: object
                properties:
                  draft:
                    $ref: "#/components/schemas/CarePlanDraft"
                required: [draft]
        "400":
          $ref: "#/components/responses/Error"
        "502":
          $ref: "#/components/responses/Error"
  /care-plans/{clientId}/versions:
    parameters:
      - in: path
        name: clientId
        required: true
        schema:
          type: string
          minLength: 1
    get:
      operationId: getCarePlanVersions
      tags: [care-plans]
      summary: List prior confirmed care plan versions
      responses:
        "200":
          description: Version history, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CarePlanVersion"
        "400":
          $ref: "#/components/responses/Error"
    post:
      operationId: confirmCarePlan
      tags: [care-plans]
      summary: Confirm and save a new care plan version
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmCarePlanRequest"
      responses:
        "201":
          description: Saved version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarePlanVersion"
        "400":
          $ref: "#/components/responses/Error"
...
  /family-updates/session:
    post:
      operationId: redeemFamilyInvite
      tags: [family-updates]
      summary: Exchange a one-time recipient invite for an HttpOnly family session
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/FamilyInviteInput" } } }
      parameters:
        - { in: header, name: X-CAREi-Family, required: true, schema: { type: string, enum: ["1"] } }
      responses:
        "200": { description: Recipient session established, content: { application/json: { schema: { $ref: "#/components/schemas/FamilyRecipient" } } } }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
    get:
      operationId: getFamilySession
      tags: [family-updates]
      security: [{ familySession: [] }]
      responses:
        "200": { description: Authenticated recipient, content: { application/json: { schema: { $ref: "#/components/schemas/FamilyRecipient" } } } }
        "401": { $ref: "#/components/responses/Error" }
    delete:
      operationId: endFamilySession
      tags: [family-updates]
      security: [{ familySession: [] }]
      parameters:
        - { in: header, name: X-CAREi-Family, required: true, schema: { type: string, enum: ["1"] } }
      responses:
        "204": { description: Session revoked }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /family-updates/consent:
    get:
      operationId: getFamilyUpdateConsent
      tags: [family-updates]
      security: [{ familySession: [] }]
      parameters:
        - { in: query, name: clientId, required: true, schema: { type: string } }
        - { in: query, name: familyMemberId, required: true, schema: { type: string } }
      responses:
        "200": { description: Consent, content: { application/json: { schema: { $ref: "#/components/schemas/FamilyUpdateConsentStatus" } } } }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
    put:
      operationId: updateFamilyUpdateConsent
      tags: [family-updates]
      security: [{ familySession: [] }]
      parameters:
        - { in: header, name: X-CAREi-Family, required: true, schema: { type: string, enum: ["1"] } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/FamilyUpdateConsentInput" } } }
      responses:
        "200": { description: Updated consent, content: { application/json: { schema: { $ref: "#/components/schemas/FamilyUpdateConsentStatus" } } } }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /family-updates:
    get:
      operationId: listFamilyUpdates
      tags: [family-updates]
      security: [{ familySession: [] }]
      parameters:
        - { in: query, name: clientId, required: true, schema: { type: string } }
        - { in: query, name: familyMemberId, required: true, schema: { type: string } }
      responses:
        "200":
          description: Sent updates, newest first
          content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/FamilyUpdateDelivery" } } } }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    familySession:
      type: apiKey
      in: cookie
      name: carei_family_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: CAREiAssistantSession
  responses:
    Error:
      description: Request failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    CarePlanSuggestion:
      type: object
      additionalProperties: false
      properties:
        suggestion:
          type: string
        source:
          type: [string, "null"]
      required: [suggestion, source]
    CarePlanDraft:
      type: object
      additionalProperties: false
      properties:
        personalDetailsAndPreferences:
          type: object
          additionalProperties: false
          properties:
            summary:
              $ref: "#/components/schemas/CarePlanSuggestion"
          required: [summary]
        needs:
          type: object
          additionalProperties: false
          properties:
            personalCare:
              $ref: "#/components/schemas/CarePlanSuggestion"
            mobility:
              $ref: "#/components/schemas/CarePlanSuggestion"
            nutrition:
              $ref: "#/components/schemas/CarePlanSuggestion"
            medication:
              $ref: "#/components/schemas/CarePlanSuggestion"
            social:
              $ref: "#/components/schemas/CarePlanSuggestion"
          required: [personalCare, mobility, nutrition, medication, social]
        identifiedRisks:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CarePlanSuggestion"
        goalsAndDesiredOutcomes:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CarePlanSuggestion"
        dailyRoutine:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CarePlanSuggestion"
        notes:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CarePlanSuggestion"
      required:
        - personalDetailsAndPreferences
        - needs
        - identifiedRisks
        - goalsAndDesiredOutcomes
        - dailyRoutine
        - notes
    CarePlanClient:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
        age:
          type: integer
          minimum: 0
        address:
          type: string
      required: [name, age, address]
    GenerateCarePlanRequest:
      type: object
      additionalProperties: false
      properties:
        client:
          $ref: "#/components/schemas/CarePlanClient"
        assessmentInput:
          type: string
          minLength: 1
          maxLength: 20000
      required: [client, assessmentInput]
    ConfirmCarePlanRequest:
      type: object
      additionalProperties: false
      properties:
        assessmentInput:
          type: string
          minLength: 1
          maxLength: 20000
        plan:
          $ref: "#/components/schemas/CarePlanDraft"
      required: [assessmentInput, plan]
    CarePlanVersion:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
        clientId:
          type: string
        version:
          type: integer
        assessmentInput:
          type: string
        plan:
          $ref: "#/components/schemas/CarePlanDraft"
        confirmedBy:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clientId
        - version
        - assessmentInput
        - plan
        - confirmedBy
        - createdAt
...
    FamilyInviteInput:
      type: object
      additionalProperties: false
      properties:
        code: { type: string, pattern: "^[A-Za-z0-9_-]{43}$" }
      required: [code]
    FamilyRecipient:
      type: object
      properties:
        clientId: { type: string }
        familyMemberId: { type: string }
        familyMemberName: { type: string }
      required: [clientId, familyMemberId, familyMemberName]
    FamilyUpdateConsentInput:
      type: object
      additionalProperties: false
      properties:
        clientId: { type: string, minLength: 1 }
        familyMemberId: { type: string, minLength: 1 }
        familyMemberName: { type: string, minLength: 1 }
        optedIn: { type: boolean }
      required: [clientId, familyMemberId, familyMemberName, optedIn]
...
    FamilyUpdateConsentStatus:
      type: object
      properties:
        clientId: { type: string }
        familyMemberId: { type: string }
        familyMemberName: { type: string }
        optedIn: { type: boolean }
        consentedAt: { type: [string, "null"], format: date-time }
        withdrawnAt: { type: [string, "null"], format: date-time }
      required: [clientId, familyMemberId, familyMemberName, optedIn, consentedAt, withdrawnAt]
    FamilyUpdateDelivery:
      type: object
      properties:
        id: { type: integer }
        visitKey: { type: string }
        clientId: { type: string }
        familyMemberId: { type: string }
        familyMemberName: { type: string }
        channel: { type: string, enum: [in_app] }
        status: { type: string, enum: [pending, processing, sent, failed, cancelled] }
        summary: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
        sentAt: { type: [string, "null"], format: date-time }
      required: [id, visitKey, clientId, familyMemberId, familyMemberName, channel, status, createdAt, sentAt]
...