# lib/api-zod/src/generated/types/confirmCarePlanRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CarePlanDraft } from "./carePlanDraft";

export interface ConfirmCarePlanRequest {
  /**
   * @minLength 1
   * @maxLength 20000
   */
  assessmentInput: string;
  plan: CarePlanDraft;
}